Python API code review fixes

# Conflicts:
#	rust/perspective-python/perspective/__init__.py
