+++ cmake/modules/FindInstallDependency.cmake
@@ -48,11 +48,16 @@ function(psp_build_dep name cmake_file)
         set(ARROW_LZ4 ON)
         set(ARROW_WITH_ZSTD ON)
         set(ARROW_WITH_LZ4 ON)
-        set(ARROW_ENABLE_THREADING OFF)
         set(ARROW_NO_EXPORT ON)
-        if(WIN32 AND NOT PSP_BUILD_WASM)
-            set(ARROW_DEPENDENCY_SOURCE "BUNDLED")
+        if(PSP_WASM_BUILD)
+            set(ARROW_ENABLE_THREADING OFF)
+        else()
+            set(ARROW_ENABLE_THREADING ON)
+            if(WIN32)
+                set(ARROW_DEPENDENCY_SOURCE "BUNDLED")
+            endif()
         endif()
+
         include_directories(SYSTEM ${CMAKE_BINARY_DIR}/${name}-build/src)
         add_subdirectory(${CMAKE_BINARY_DIR}/${name}-src/cpp/
             ${CMAKE_BINARY_DIR}/${name}-build
