+++ rust/perspective-python/build.mjs
@@ -13,6 +13,7 @@
 import * as fs from "node:fs";
 import sh from "../../tools/perspective-scripts/sh.mjs";
 import * as url from "url";
+import * as TOML from "@iarna/toml";
 
 const __dirname = url.fileURLToPath(new URL(".", import.meta.url)).slice(0, -1);
 const pkg = JSON.parse(
@@ -89,7 +90,13 @@ if (build_wheel) {
     cmd.sh(`maturin build ${flags} --features=external-cpp ${target}`);
 }
 
+const old = fs.readFileSync("./pyproject.toml");
+
 if (build_sdist) {
+    const toml = TOML.parse(old);
+    console.log(toml);
+    delete toml.tool.maturin["data"];
+    fs.writeFileSync("./pyproject.toml", TOML.stringify(toml));
     cmd.sh(`maturin sdist`);
 }
 
@@ -98,3 +105,7 @@ if (!build_wheel && !build_sdist) {
 }
 
 cmd.runSync();
+
+if (build_sdist) {
+    fs.writeFileSync("./pyproject.toml", old);
+}
