+++ rust/perspective-python/perspective/handlers/__init__.py
@@ -9,18 +9,3 @@
 #  ┃ This file is part of the Perspective library, distributed under the terms ┃
 #  ┃ of the [Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0). ┃
 #  ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-
-try:
-    from .aiohttp import PerspectiveAIOHTTPHandler  # noqa: F401
-except ImportError:
-    ...
-
-try:
-    from .starlette import PerspectiveStarletteHandler  # noqa: F401
-except ImportError:
-    ...
-
-try:
-    from .tornado import PerspectiveTornadoHandler  # noqa: F401
-except ImportError:
-    ...
