+++ rust/perspective-python/perspective/handlers/starlette.py
@@ -11,6 +11,7 @@
 #  ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
 
 import asyncio
+import perspective
 
 
 class PerspectiveStarletteHandler(object):
@@ -31,7 +32,7 @@ class PerspectiveStarletteHandler(object):
     """
 
     def __init__(self, **kwargs):
-        self._server = kwargs.pop("perspective_server")
+        self._server = kwargs.pop("perspective_server", perspective.GLOBAL_SERVER)
         self._websocket = kwargs.pop("websocket")
         super().__init__(**kwargs)
