+++ rust/perspective-python/pyproject.toml
@@ -18,6 +18,7 @@ build-backend = "maturin"
 name = "perspective-python"
 requires-python = ">=3.9"
 dynamic = ["version"]
+dependencies = []
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Programming Language :: Rust",
@@ -33,10 +34,12 @@ classifiers = [
     # "Framework :: Jupyter :: JupyterLab :: Extensions :: Mime Renderers",
     "Framework :: Jupyter :: JupyterLab :: Extensions :: Prebuilt",
 ]
-dependencies = []
-[project.optional-dependencies]
-jupyter = ["Jinja2>=2.0,<4", "ipywidgets>=7.5.1,<=8.1.3"]
 
+[project.optional-dependencies]
+jupyter = ["ipywidgets>=7.5.1,<9"]
+tornado = ["tornado>=5,<7"]
+aiohttp = ["aiohttp>=2,<4"]
+starlette = ["starlette<1"]
 
 [tool.maturin]
 module-name = "perspective"
