+++ rust/perspective-python/requirements.txt
@@ -2,7 +2,6 @@ pytest>=7.4.3
 ruff>=0.4.6
 Faker==26.0.0
 ipywidgets==8.1.3
-Jinja2==3.1.4
 jupyterlab==4.2.3
 jupyterlab_git==0.50.1
 maturin==1.6.0
@@ -12,6 +11,7 @@ pandas==2.2.2
 pyarrow==16.1.0
 psutil==6.0.0
 pytest==8.2.2
+pytest-sugar==1.0.0
 ruff==0.5.0
 tornado==6.4.1
 traitlets==5.14.3
